Finished PUT

+++ src/app/components/admin-page/admin-page.component.ts
@@ -46,36 +46,37 @@ export class AdminPageComponent implements OnInit {
   }
 
   deleteOffers(id: any) {
-    this.delete(id);
-    this.offersService.deleteOffers(id).subscribe((data: any) => {
-      this.getOffers();
-    });
+    if(confirm("Are you sure to delete ")) {
+      console.log("Implement delete functionality here");
+      this.offersService.deleteOffers(id).subscribe((data: any) => {
+        this.getOffers();
+      });
+    }
+    
 
   }
 
-  delete(id:number){
-    console.log(id);
-  }
-
-  edit(item: any){
-    console.log(item);
-  }
 
   openAddOffer(data: any){
-    this.edit(data);
+
    this.modalService.openAddOffer(data).then((data) => {
     if(data) {
-      // PUT PREMA BE
-      /*this.contentService.editContent(data, contentData.id).subscribe(() => {
-        this.getContent();
-        this.isEdit = false;
-      });
-      */
+      let post = {
+        name: data.name,
+        price: data.price,
+        amount: data.amount,
+        media: data.media,
+      }
+      this.offersService.editOffer(post, data.id).subscribe((data: any)=> {
+        console.log("Successfully updated offer");
+        this.getOffers();
+      } );
+     
     }
+    
   })
   .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
 
-      
     }
   
     logout(){
